add ricerca per autore method

+++ src/main/java/giovanni/exception/NotFoundException.java
@@ -1,7 +1,7 @@
 package giovanni.exception;
 
 public class NotFoundException extends RuntimeException {
-    public NotFoundException(long id) {
+    public NotFoundException(int id) {
         super("Il record con l'id " + id + " non è stato trovato!");
     }
 }
